Make the code more modular
Starting with logger
